pyproject.toml
[build-system]
requires = ["setuptools>=65.5.1", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clud"
readme = "README.md"
description = "Claude Code in YOLO mode"
requires-python = ">=3.10"
keywords = ["claude", "cli", "ai"]
license = { text = "BSD 3-Clause License" }
classifiers = ["Programming Language :: Python :: 3"]
dependencies = [
    "keyring>=24.0.0",
    "fasteners>=0.17.0",
    "appdirs>=1.4.4",
    "disklru>=0.1.0",
    "httpx>=0.24.0",
    "pywinpty>=3.0.0; sys_platform == 'win32'",
    "running-process>=1.0.7",
    "codeup>=1.0.8",
    "nodeenv>=1.8.0",
    "nodejs-wheel>=22.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "markdown2>=2.4.0",
    "websockets>=12.0",
    # Cluster control plane dependencies
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "psutil>=5.9.0",
    "python-telegram-bot>=20.0",
]
# Change this with the version number bump.
version = "1.0.22"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-xdist>=2.0.0",
    "ruff>=0.1.0",
    "pyright>=1.1.0",
]
keyring-fallbacks = [
    "keyrings.cryptfile>=1.3.0",
    "cryptography>=3.0.0",
]
posix = [
    "zcmds_win32>=1.2.0; sys_platform == 'win32'",
]
full = [
    "zcmds_win32>=1.2.0; sys_platform == 'win32'",
    # "clud-cluster>=0.1.0",  # Temporarily commented for Phase 1 development
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.ruff]
line-length = 200
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "C4", "ANN"]
ignore = [
    "B008",  # Allow function calls in argument defaults (FastAPI Depends pattern)
    "ANN401",  # Allow typing.Any when necessary (e.g., __exit__ parameters)
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
force-single-line = false
split-on-trailing-comma = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
addopts = "--durations=0"

[tool.pyright]
include = ["src", "tests"]
exclude = ["**/__pycache__"]

# Type checking configuration
typeCheckingMode = "strict"
pythonVersion = "3.10"
pythonPlatform = "All"
useLibraryCodeForTypes = true
analyzeUnannotatedFunctions = true
strictParameterNoneValue = true
strictListInference = true
strictDictionaryInference = true
strictSetInference = true

# Report settings
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = false
reportUnusedClass = false
reportUnusedFunction = false
reportUnusedVariable = false
reportDuplicateImport = true
reportWildcardImportFromLibrary = true
reportOptionalSubscript = false
reportOptionalMemberAccess = false
reportOptionalCall = false
reportOptionalIterable = false
reportOptionalContextManager = false
reportOptionalOperand = false
reportGeneralTypeIssues = true
reportPropertyTypeMismatch = true
reportFunctionMemberAccess = true
reportPrivateUsage = false
reportConstantRedefinition = false
reportIncompatibleMethodOverride = true
reportIncompatibleVariableOverride = true
reportInconsistentConstructor = true
reportOverlappingOverload = true
reportMissingSuperCall = false
reportUninitializedInstanceVariable = false
reportInvalidStringEscapeSequence = true
reportUnknownParameterType = "error"
reportUnknownArgumentType = "error"
reportUnknownLambdaType = "error"
reportUnknownVariableType = "error"
reportUnknownMemberType = "error"
reportMissingParameterType = "error"
reportMissingTypeArgument = "error"
reportMissingReturnType = "error"
reportUntypedFunctionDecorator = "error"
reportUntypedClassDecorator = "error"
reportUntypedBaseClass = "error"
reportUntypedNamedTuple = "error"
reportInvalidTypeVarUse = true
reportCallInDefaultInitializer = false
reportUnnecessaryIsInstance = false
reportUnnecessaryCast = false
reportUnnecessaryComparison = false
reportUnnecessaryContains = false
reportAssertAlwaysTrue = false
reportSelfClsParameterName = true
reportImplicitStringConcatenation = false
reportUndefinedVariable = true
reportUnboundVariable = true
reportInvalidStubStatement = true
reportIncompleteStub = true
reportUnsupportedDunderAll = true
reportUnusedCallResult = false
reportUnusedCoroutine = true
reportUnusedExpression = false
reportUnnecessaryTypeIgnoreComment = false
reportMatchNotExhaustive = true

# Display settings
verboseOutput = false
autoImportCompletions = false
indexing = false
functionSignatureDisplay = "compact"

[project.scripts]
clud = "clud.cli:main"
clud-cluster = "clud.cluster.cli:main"
